fixed potential bug

+++ pfg-termanal/multi_term.sh
@@ -3,5 +3,7 @@
 for file in $1/*.pdb
 do
   python3 pfg_run.py --p $file --o $2
-  rm termanal_support
+  if [ -f termanal_support ]; then
+    rm termanal_support
+  fi
 done
